Rename `find_matches_and_compute_last` into `find_matches_and_compute_f7` for consistency
