+++ crates/farmer/ab-proof-of-space-gpu/src/shader/find_proofs.rs
@@ -6,7 +6,7 @@ mod gpu_tests;
 use crate::shader::constants::{
     K, NUM_MATCH_BUCKETS, NUM_S_BUCKETS, NUM_TABLES, REDUCED_MATCHES_COUNT,
 };
-use crate::shader::find_matches_and_compute_last::NUM_ELEMENTS_PER_S_BUCKET;
+use crate::shader::find_matches_and_compute_f7::NUM_ELEMENTS_PER_S_BUCKET;
 use crate::shader::types::{Position, PositionExt};
 use core::mem::MaybeUninit;
 use spirv_std::arch::{
